api/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
elasticsearch==7.8.1
environs==8.0.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
neo4j-driver==4.1.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
Werkzeug==1.0.1
